+++ Procfile-backends
@@ -1,2 +1,2 @@
-beat: celery beat -A mergify_engine.worker
-worker: celery worker --pool gevent -A mergify_engine.worker --task-events -Q mergify,celery
+beat: celery beat -A mergify_engine.tasks
+worker: celery worker --pool gevent -A mergify_engine.tasks --task-events -Q mergify,celery
