+++ mergify_engine/tests/functional/test_engine.py
@@ -25,6 +25,7 @@ from mergify_engine import config
 from mergify_engine import context
 from mergify_engine import debug
 from mergify_engine import engine
+from mergify_engine import tasks
 from mergify_engine.clients import github
 from mergify_engine.tests.functional import base
 
@@ -33,11 +34,7 @@ LOG = logging.getLogger(__name__)
 
 
 def run_smart_strict_workflow_periodic_task():
-    # NOTE(sileht): actions must not be loaded manually before the celery
-    # worker. Otherwise we have circular import loop.
-    from mergify_engine.actions.merge import queue
-
-    queue.smart_strict_workflow_periodic_task.apply_async()
+    tasks.smart_strict_workflow_periodic_task.apply_async()
 
 
 class TestEngineV2Scenario(base.FunctionalTestBase):
